+++ api/adm_changeUserState.php
@@ -26,8 +26,15 @@
             $decoded = JWT::decode($jwt, $key, array('HS256'));
             $user = ORM::for_table('users')->find_one($_GET['uID']);
             if($user) {
-                
-                echo "success";
+                if($_GET['action']=='disable') {
+                    $user->accState=0;
+                    $user->save();
+                    echo "success";
+                } elseif($_GET['action']=='enable') {
+                    $user->accState=1;
+                    $user->save();
+                    echo "success";
+                }                
             } else {
                 echo "Error: user with id ".$_GET['uID']." not found!";
             }